diag: PANDDA_MEMTRACE per-stage memory trace + array_load float32
Add a PANDDA_MEMTRACE-gated trace logging system-wide used/available RAM (what
the OOM killer sees, incl. Ray workers + plasma) and process RSS at each
per-dataset stage: dataset_start, dmaps_built, unmask_xmap, characterization,
filter_models, process_models_done, autobuild_start/done. Flushed per line so
the last line survives a SIGKILL -- to pinpoint the large-cell OOM stage, since
no static allocation (comparator maps 27 MB f32, dmaps <=3 GB, masks Int8)
explains the 128 GB kill.

Also: array_load allocated its (datasets x voxels) matrix as np.zeros default
float64 while the maps are float32 -- make it float32 (halves it).

+++ pandda_gemmi/dmaps/sparse_dmap_stream.py
@@ -90,8 +90,9 @@ class SparseDMapStream:
         # Get the shape to load datasets into
         shape = (len(self.datasets), self.dframe.mask.indicies[0].size)
 
-        # Get the array
-        array = np.zeros(shape)
+        # Get the array (float32 -- the maps are f32; np.zeros defaults to f64
+        # and would gratuitously double this datasets x voxels matrix).
+        array = np.zeros(shape, dtype=np.float32)
 
         # Load each dataset in
         for j, dtag in enumerate(self.datasets):
